Log in, reading and sending messages (Auth, Database, Storage)
User is required to sign in with Google (Auth), then on the
MainActivity, the messages stored in the real-time database
(Database) are displayed. The user can push a new message to the
real-time database, which will be automatically synced and
displayed locally by the FirebaseRecyclerAdapter.

The user can also send an image message. In this case, a message
with an image URL is saved in the real-time database, and the
image itself is uploaded to the Firebase storage (Storage).